paginator/Paginator.php creation
- Remove paginator() and displayPage() functions from Manager.php
- Create paginator/Paginator.php and add paginator() and displayPage() functions in it.
- Modify Manager.php to add a Paginator object creation in __construct function
- Modify PostManager.php to use paginator object.
- Minor modifications in Post.php
- Modify PublicController.php to remove call to Paginator functions (add to PostManager Functions)
