+++ model/Manager.php
@@ -2,33 +2,15 @@
 
 namespace EmilieSchott\BlogPHP\Model;
 
+use EmilieSchott\BlogPHP\Paginator\Paginator;
 use EmilieSchott\BlogPHP\Model\PDOFactory;
 
-abstract class Manager {
+abstract class Manager extends Paginator {
     protected $db;
+    protected $paginator;
     
     public function __construct() {
-        $this->db=PDOFactory::getDbConnection();
-    }
-
-    public function paginator(array $datas, int $itemsPerPage): array {
-        $datasPages = \array_chunk($datas, $itemsPerPage);
-        \array_unshift($datasPages, '');
-        unset($datasPages[0]);
-        $pagesNbr = \array_key_last($datasPages);
-
-        $paginator = [
-        'datasPages' => $datasPages,
-        'pagesNbr' => $pagesNbr
-        ];
-
-        return $paginator;
-    }
-
-    public function displayPage(array $datasPages, int $page): array {
-        $offset=$page-1;
-        $datasPage=array_slice($datasPages, $offset, 1);
-        $datasPage=$datasPage[0];
-        return $datasPage;
+        $this->db = PDOFactory::getDbConnection();
+        $this->paginator = new Paginator(); 
     }
 }
